workers/requirements.txt
psycopg2-binary>=2.9
redis>=5.0
minio>=7.2
requests>=2.31
apscheduler>=3.10
python-dotenv>=1.0
shapely>=2.0
geojson>=3.1
rasterio>=1.3
numpy>=1.26
Pillow>=10.2
